+++ R/token.R
@@ -13,7 +13,7 @@
 #' @param version The AAD version, either 1 or 2. Authenticating with a personal account as opposed to a work or school account requires AAD 2.0.
 #' @param authorize_args An optional list of further parameters for the AAD authorization endpoint. These will be included in the request URI as query parameters. Only used if `auth_type="authorization_code"`.
 #' @param token_args An optional list of further parameters for the token endpoint. These will be included in the body of the request for `get_azure_token`, or as URI query parameters for `get_managed_token`.
-#' @param use_cache If TRUE and cached credentials exist, use them instead of obtaining a new token. Set this to FALSE to bypass the cache.
+#' @param use_cache If TRUE and cached credentials exist, use them instead of obtaining a new token. The default value of NULL means to use the cache only if AzureAuth is not running inside a Shiny app.
 #' @param on_behalf_of For the on-behalf-of authentication type, a token. This should be either an AzureToken object, or a string containing the JWT-encoded token itself.
 #' @param auth_code For the `authorization_code` flow, the code. Only used if `auth_type == "authorization_code"`.
 #' @param device_creds For the `device_code` flow, the device credentials used to verify the session between the client and the server. Only used if `auth_type == "device_code"`.
@@ -224,7 +224,7 @@
 get_azure_token <- function(resource, tenant, app, password=NULL, username=NULL, certificate=NULL, auth_type=NULL,
                             aad_host="https://login.microsoftonline.com/", version=1,
                             authorize_args=list(), token_args=list(),
-                            use_cache=TRUE, on_behalf_of=NULL, auth_code=NULL, device_creds=NULL)
+                            use_cache=NULL, on_behalf_of=NULL, auth_code=NULL, device_creds=NULL)
 {
     auth_type <- select_auth_type(password, username, certificate, auth_type, on_behalf_of)