+++ NEWS.md
@@ -1,5 +1,6 @@
 # AzureAuth 1.3.1.9000
 
+- Change the default caching behaviour to disable the cache if running inside Shiny.
 - Update Shiny vignette to clean up redirect page after authenticating (thanks to Tyler Littlefield).
 
 # AzureAuth 1.3.1